+++ apm-agent-plugins/apm-mongoclient-plugin/apm-mongoclient-plugin-3_4/pom.xml
@@ -50,13 +50,7 @@
             <groupId>org.mongodb</groupId>
             <artifactId>mongodb-driver-async</artifactId>
             <version>${version.mongo}</version>
-            <scope>provided</scope>
+            <scope>test</scope>
         </dependency>
-        <!--<dependency>-->
-            <!--<groupId>de.flapdoodle.embed</groupId>-->
-            <!--<artifactId>de.flapdoodle.embed.mongo</artifactId>-->
-            <!--<version>1.50.5</version>-->
-            <!--<scope>test</scope>-->
-        <!--</dependency>-->
     </dependencies>
 </project>
