+++ apm-agent-plugins/apm-mongoclient-plugin/apm-mongoclient-plugin-3_4/src/main/resources/META-INF/services/co.elastic.apm.agent.bci.ElasticApmInstrumentation
@@ -1,2 +1,1 @@
-co.elastic.apm.agent.mongoclient.v3_4.MongoClientSyncInstrumentation
-co.elastic.apm.agent.mongoclient.v3_4.MongoClientAsyncInstrumentation
+co.elastic.apm.agent.mongoclient.v3_4.MongoClientCommonInstrumentation
